refactor: 뒤로가기 url 커스텀훅으로 분리
